init logger in main func as soon as possible
- add logger factory
 It is difficult to test it. Only way how to test logger is to write into []bytes and compare. Such tests are difficult for maintain.

in order to move depresolver at the very beginning of the main function I had to move client.Client parameter from depresolver constructor fonc do ResolveSpec() func. The reason for that is to build log as soon as possible in the main function

resolving configuration is the first step of bootstrap. The second step is creating logger instance. Thanks to that, I removed the redundant setuploger instance.